Update useEffect() at Chat.js
Import functions from Firebase.  Create the onSnapshot() listener on a query that targets the messages collection in useEffect(). Use orderBy function and createdAt property to sort the query results in descending order. Construct an array of messages from the fetched documents and assign this array to the messages state using setMessages() inside of the callback onSnapshot().
Convert the TimeStamp stored at the createdAt property of each message to a Date object that Gifted Chat understands. Call the unsubMessages function.