Update CustomActions with useActionSheet
Import useActionSheet. Create function called onActionPress to display action menu "ActionSheet" that contains options. Determine the position of cancel button of ActionSheet, so that ActionSheet can close the view if the user presses "Cancel". Use showActionSheetWithOptions() function to initialize and show the ActionSheet. Call the onActionPress function in case of btn press.